+++ controller/server/gui.mli
@@ -1,4 +1,9 @@
-val routes :
-  connman : Connman.Manager.t
+val start :
+  port : int
+  -> shutdown : (unit -> unit Lwt.t)
+  -> health_s : Health.state Lwt_react.S.t
+  -> update_s : Update.state Lwt_react.S.t
+  -> rauc : Rauc.t
+  -> connman : Connman.Manager.t
   -> internet: Network.Internet.state Lwt_react.S.t
-  -> Opium.App.builder
+  -> unit Lwt.t