+++ system/configuration.nix
@@ -105,18 +105,4 @@ with lib;
     group = "root";
   };
 
-  # Start controller
-  systemd.services.playos-controller = {
-    description = "PlayOS Controller";
-    serviceConfig = {
-      ExecStart = "${pkgs.playos-controller}/bin/playos-controller ${config.playos.updateUrl}";
-      User = "root";
-      RestartSec = "10s";
-      Restart = "always";
-    };
-    wantedBy = [ "multi-user.target" ];
-    requires = [ "rauc" "connman" ];
-    after = [ "rauc" "connman" ];
-  };
-
 }