+++ Changelog.md
@@ -1,3 +1,13 @@
+# [2019.2.5-beta] - 2019-02-19
+
+## Added
+
+- controller: print labels
+
+## Changed
+
+- controller: show more information and add ability to remove service
+
 # [2019.2.4-beta] - 2019-02-15
 
 ## Changed